+++ tests/s2f_recipe/test_estimate_bouton_reduction.py
@@ -5,10 +5,9 @@ import numpy as np
 import numpy.testing as npt
 from bluepy import Circuit
 from mock import MagicMock, patch
+from utils import TEST_DATA_DIR
 
 import connectome_tools.s2f_recipe.estimate_bouton_reduction as test_module
-
-TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "../data")
 
 
 @patch(test_module.__name__ + ".sample_bouton_density", return_value=np.array([1.0, 3.0]))