+++ tests/test_dataset.py
@@ -5,10 +5,9 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 from parameterized import param, parameterized
+from utils import TEST_DATA_DIR
 
 from connectome_tools.dataset import _remove_duplicates, read_bouton_density, read_nsyn
-
-TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
 
 
 @parameterized.expand(